# updated bigint with optimized mul and from_big_indian