--- a/util/bigint/Cargo.toml
+++ b/util/bigint/Cargo.toml
@@ -8,7 +8,7 @@ version = "0.1.3"
 authors = ["Parity Technologies <admin@parity.io>"]
 
 [dependencies]
-bigint = "3.0"
+bigint = "4.0"
 rustc-hex = "1.0"
 rand = "0.3.12"
 libc = "0.2"