bigint optimizations from https://github.com/paritytech/bigint/pull/23

reading u256 is now 60% faster

old rlp benches:

```rust
test bench_decode_u256_value         ... bench:         118 ns/iter (+/- 7)
test bench_decode_u64_value          ... bench:          77 ns/iter (+/- 18)
```

new benches:

```rust
test bench_decode_u256_value         ... bench:          73 ns/iter (+/- 5)
test bench_decode_u64_value          ... bench:          80 ns/iter (+/- 32)
```